snapcraft.yaml: fix Intel hardware acceleration
The OpenCL ICD provided by Intel hard-codes the path to the OpenCL
library. This works in more traditional use-cases, but because
libOpenCL.so is NOT in the usual path here, the usual LD_LIBRARY_PATH
workarounds don't necessarily apply. Instead, spoof the location of
libOpenCL.so similarly to how spoofing the ICD file itself is done (via
a layout).

Additionally, enforce using libraries provided by jellyfin-ffmpeg to
ensure configure consistency for codecs and to improve the available
features when using ffmpeg.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>