snapcraft.yaml: add 'read:all' attribute to home plug
As the jellyfin service runs as root, the home plug only grants
access to /root. As the intention of the plug is to gain access to
/home/<user>, instead the `read: all` attribute must be used. This
enables the jellyfin service to access every directory in /home/foo.

This of course opens up the opportunity for the snap to view *every*
user's home directory, which may be undesirable. But there is
(currently) no good method for specifying only a particular user's home
directory.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>