snap: implement check-health functionality
This feature will act to "notify" users if something is wrong with
their Jellyfin instance and attempt to identify why. The feature set of
'snap health' is rudimentary and the onus is heavily on the publisher to
identify what can fail, why, and report that to the user appropriately.

There is currently no way for a publisher to know if a snap deployed on
any given system is unhealthy; it's a feature gap in snapd which has yet
to be implemented.

The check-health hook should be run every 30 minutes if the health is
"okay", and every five minutes otherwise. The hook as written should not
substantially impact performance in either case.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>